Feature: added profile icon to NavigationBar
