+++ src/Components/NavigationBar.js
@@ -1,26 +1,91 @@
-import "bootstrap/dist/css/bootstrap.css";
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
-import Navbar from 'react-bootstrap/Navbar'
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import Menu from '@mui/material/Menu';
+import Container from '@mui/material/Container';
+import Avatar from '@mui/material/Avatar';
+import Tooltip from '@mui/material/Tooltip';
+import MenuItem from '@mui/material/MenuItem';
 import Logo from './Logo'
-import Container from 'react-bootstrap/Container'
+import AccountCircleRoundedIcon from '@mui/icons-material/AccountCircleRounded';
 
-function NavigationBar(){
-    return(
-        <Container>
-            <Navbar className='navigationBar' expand='true' fixed='top' style={{margin: 0, paddingBottom: '1%'}}>
-                <Logo/>
-            </Navbar>
-        </Container>
-        /*
-        <Row className="navigationBar">
-            <div className="NavigationBar" style={{}}>
+//const pages = [];
+const settings = ['Account', 'Logout'];
 
-            </div>
+const NavigationBar = () => {
+  const [anchorElUser, setAnchorElUser] = React.useState(null);
 
-        </Row>
-        */
-    );
-}
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
 
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+
+  return (
+    <AppBar position="static" sx={{backgroundColor:'white'}}>
+      <Container maxWidth="xl">
+        <Toolbar disableGutters>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ mr: 2, display: { xs: 'none', md: 'flex' } }}
+          >
+            <Logo/>
+          </Typography>
+
+          <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'right' } }}>
+            <IconButton
+              size="small"
+              aria-label="account of current user"
+              aria-controls="menu-appbar"
+              aria-haspopup="true"
+              color="inherit"
+            >
+            </IconButton>
+         </Box>
+
+          <Box sx={{ flexGrow: 0 }}>
+            <Tooltip title="Open settings">
+              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
+                <AccountCircleRoundedIcon fontSize='large' color='EFF0F3'/>
+              </IconButton>
+            </Tooltip>
+            <Menu
+              sx={{ mt: '45px' }}
+              id="menu-appbar"
+              anchorEl={anchorElUser}
+              anchorOrigin={{
+                vertical: 'top',
+                horizontal: 'right',
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: 'top',
+                horizontal: 'right',
+              }}
+              open={Boolean(anchorElUser)}
+              onClose={handleCloseUserMenu}
+            >
+              {settings.map((setting) => (
+                <MenuItem key={setting} onClick={handleCloseUserMenu}>
+                  <Typography textAlign="center">{setting}</Typography>
+                </MenuItem>
+              ))}
+            </Menu>
+          </Box>
+        </Toolbar>
+      </Container>
+    </AppBar>
+  );
+};
+
+/**
+ * <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
+ */
 export default NavigationBar;
